feat: Summarize the manually curated matches
The matches were summarized for their counts and item types in
analysis/analyse_matches.py, this commit also accompanies the
documentation and associated draft for this analysis.

The wdt module also changed name to be the quality module, since it
now contains all functions related to item quality, such as summarizing
and plotting.